refactor: name single-export type modules after their export
reportView -> timeRangePickerConfig, shared/reporting -> reportingBasis, shared/s3 ->
s3PresignedUrl, shared/viewport -> view, shared/toast -> toastData, utility/promises ->
awaitable, and document the rule in the types skill.
