+++ src/components/ui/DatePickers/DatePicker/DatePickerInput.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from 'react'
 import { TriangleAlert } from 'lucide-react'
 
-import type { View } from '@internal-types/shared/viewport'
+import type { View } from '@internal-types/shared/view'
 import { DateInput, DateSegment } from '@ui/Date/Date'
 import { FieldError } from '@ui/Form/Form'
 import { InputGroup } from '@ui/Input/InputGroup'
