+++ src/hooks/features/linkedAccounts/usePollPlaidHostedLinkStatus.ts
@@ -1,6 +1,6 @@
 import { useCallback, useEffect } from 'react'
 
-import type { Awaitable } from '@internal-types/utility/promises'
+import type { Awaitable } from '@internal-types/utility/awaitable'
 import { type ApiPlaidHostedLinkStatus, PlaidHostedLinkState } from '@schemas/linkedAccounts/plaidHostedLinkStatus'
 import { APIError } from '@utils/shared/api/apiError'
 import { useGetPlaidHostedLinkStatus } from '@api/businesses/[business-id]/plaid/hosted-link/get'
