+++ src/hooks/features/reports/useResolvedReportView.ts
@@ -1,4 +1,4 @@
-import { type View } from '@internal-types/shared/viewport'
+import { type View } from '@internal-types/shared/view'
 import { useElementViewSize } from '@hooks/utils/size/useElementViewSize'
 
 export const useResolvedReportView = (passedView?: View) => {
