+++ src/hooks/legacy/useUpdateOpeningBalanceAndDate.ts
@@ -1,6 +1,6 @@
 import useSWRMutation from 'swr/mutation'
 
-import type { Awaitable } from '@internal-types/utility/promises'
+import type { Awaitable } from '@internal-types/utility/awaitable'
 import { post } from '@utils/shared/api/authenticatedHttp'
 import { createBuildKey } from '@utils/shared/swr/createBuildKey'
 import { useBuildKeyInputs } from '@hooks/utils/swr/useBuildKeyInputs'
