+++ src/providers/global/LayerContext/BusinessProvider.tsx
@@ -8,7 +8,7 @@ import {
   type LayerContextValues,
   type LayerThemeConfig,
 } from '@internal-types/shared/layerContext'
-import { type ToastData } from '@internal-types/shared/toast'
+import { type ToastData } from '@internal-types/shared/toastData'
 import { errorHandler, type LayerError } from '@utils/shared/api/errorHandler'
 import { buildColorsPalette } from '@utils/shared/styles/colors'
 import { useGetAccountingConfiguration } from '@api/businesses/[business-id]/accounting-config/get'
