+++ src/providers/linkedAccounts/LinkedAccounts/LinkedAccountsProvider.tsx
@@ -1,6 +1,6 @@
 import { type PropsWithChildren } from 'react'
 
-import type { Awaitable } from '@internal-types/utility/promises'
+import type { Awaitable } from '@internal-types/utility/awaitable'
 import { type PlaidHostedLinkConfig } from '@schemas/linkedAccounts/plaidHostedLinkConfig'
 import { useLinkedAccounts } from '@hooks/legacy/useLinkedAccounts'
 import { LinkedAccountsContext } from '@providers/linkedAccounts/LinkedAccounts/LinkedAccountsContext'
