+++ src/types/features/bankTransactions/bankTransaction.ts
@@ -1,4 +1,4 @@
-import { type S3PresignedUrl } from '@internal-types/shared/s3'
+import { type S3PresignedUrl } from '@internal-types/shared/s3PresignedUrl'
 import {
   type BankTransactionSchema,
 } from '@schemas/bankTransactions/bankTransaction'
