+++ src/types/shared/s3PresignedUrl.ts
