+++ src/types/shared/view.ts
